Tabs and some re-organization for export
