--- Assets/FreeCamera.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: cce3bf078b091f643b81000c4c6307b2
-folderAsset: yes
-timeCreated: 1484502038
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
