+++ Assets/GunTurrets/Demo/Scripts/WobblyThing.cs
@@ -1,12 +1,15 @@
 ﻿using UnityEngine;
 
-public class WobblyThing : MonoBehaviour
+namespace TurretDemo
 {
-	public float strength;
+   public class WobblyThing : MonoBehaviour
+   {
+      public float strength;
 
-	private void Update()
-	{
-		transform.Translate(new Vector3(Mathf.Sin(Time.time), 0.0f, 0.0f) * strength * Time.deltaTime);
-	}
+      private void Update()
+      {
+         transform.Translate(new Vector3(Mathf.Sin(Time.time), 0.0f, 0.0f) * strength * Time.deltaTime);
+      }
 
+   }
 }
